# openapi.json
{"openapi":"3.0.3","info":{"title":"Api Documentation","description":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"},"version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Inferred Url"}],"tags":[{"name":"Managing Countries Api Controller"},{"name":"Managing Customer Address Api Controller"},{"name":"Managing Province Api Controller"},{"name":"basic-error-controller","description":"Basic Error Controller"}],"paths":{"/address":{"get":{"tags":["address-controller-impl"],"summary":"listCustomerAddresses","description":"List all customer addresses","operationId":"listCustomerAddressesUsingGET","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO«List«CustomerAddress»»"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["address-controller-impl"],"summary":"createAddress","description":"Create New Address","operationId":"createAddressUsingPOST","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Address"}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO«CustomerAddress»"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/address/{addressId}":{"get":{"tags":["address-controller-impl"],"summary":"findAddress","description":"Find address","operationId":"findAddressUsingGET","parameters":[{"name":"addressId","in":"path","description":"addressId","required":true,"style":"simple","schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO«CustomerAddress»"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["address-controller-impl"],"summary":"updateAddress","description":"Create New Address","operationId":"updateAddressUsingPUT","parameters":[{"name":"addressId","in":"path","description":"addressId","required":true,"style":"simple","schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Address"}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO«CustomerAddress»"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["address-controller-impl"],"summary":"deleteCustomerAddress","description":"Delete customer addresses","operationId":"deleteCustomerAddressUsingDELETE","parameters":[{"name":"addressId","in":"path","description":"addressId","required":true,"style":"simple","schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO«int»"}}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/error":{"get":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingGET","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingPUT","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingPOST","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingDELETE","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"options":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingOPTIONS","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"head":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingHEAD","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"patch":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingPATCH","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"trace":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingTRACE","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/country/list":{"get":{"tags":["country-controller-impl"],"summary":"listCountries","description":"List Countries alphabetically","operationId":"listCountriesUsingGET","parameters":[{"name":"sort","in":"query","description":"sort","required":false,"style":"form","schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO«List«Country»»"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/province/list":{"get":{"tags":["province-controller-impl"],"summary":"listProvinces","description":"Find provinces for specified country","operationId":"listProvincesUsingGET","parameters":[{"name":"countryCode","in":"path","description":"countryCode","required":true,"style":"simple","schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDTO«List«Province»»"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"components":{"schemas":{"Address":{"title":"Address","type":"object","properties":{"city":{"type":"string"},"countryCode":{"type":"string"},"line1":{"type":"string"},"line2":{"type":"string"},"provinceCode":{"type":"string"},"suburb":{"type":"string"}}},"Country":{"title":"Country","type":"object","properties":{"countryCode":{"type":"string"},"countryName":{"type":"string"}}},"CustomerAddress":{"title":"CustomerAddress","type":"object","properties":{"addressId":{"type":"integer","format":"int32"},"city":{"type":"string"},"country":{"$ref":"#/components/schemas/Country"},"line1":{"type":"string"},"line2":{"type":"string"},"province":{"$ref":"#/components/schemas/Province"},"suburb":{"type":"string"}}},"ModelAndView":{"title":"ModelAndView","type":"object","properties":{"empty":{"type":"boolean"},"model":{"type":"object"},"modelMap":{"type":"object","additionalProperties":{"type":"object"}},"reference":{"type":"boolean"},"status":{"type":"string","enum":["ACCEPTED","ALREADY_REPORTED","BAD_GATEWAY","BAD_REQUEST","BANDWIDTH_LIMIT_EXCEEDED","CHECKPOINT","CONFLICT","CONTINUE","CREATED","DESTINATION_LOCKED","EXPECTATION_FAILED","FAILED_DEPENDENCY","FORBIDDEN","FOUND","GATEWAY_TIMEOUT","GONE","HTTP_VERSION_NOT_SUPPORTED","IM_USED","INSUFFICIENT_SPACE_ON_RESOURCE","INSUFFICIENT_STORAGE","INTERNAL_SERVER_ERROR","I_AM_A_TEAPOT","LENGTH_REQUIRED","LOCKED","LOOP_DETECTED","METHOD_FAILURE","METHOD_NOT_ALLOWED","MOVED_PERMANENTLY","MOVED_TEMPORARILY","MULTIPLE_CHOICES","MULTI_STATUS","NETWORK_AUTHENTICATION_REQUIRED","NON_AUTHORITATIVE_INFORMATION","NOT_ACCEPTABLE","NOT_EXTENDED","NOT_FOUND","NOT_IMPLEMENTED","NOT_MODIFIED","NO_CONTENT","OK","PARTIAL_CONTENT","PAYLOAD_TOO_LARGE","PAYMENT_REQUIRED","PERMANENT_REDIRECT","PRECONDITION_FAILED","PRECONDITION_REQUIRED","PROCESSING","PROXY_AUTHENTICATION_REQUIRED","REQUESTED_RANGE_NOT_SATISFIABLE","REQUEST_ENTITY_TOO_LARGE","REQUEST_HEADER_FIELDS_TOO_LARGE","REQUEST_TIMEOUT","REQUEST_URI_TOO_LONG","RESET_CONTENT","SEE_OTHER","SERVICE_UNAVAILABLE","SWITCHING_PROTOCOLS","TEMPORARY_REDIRECT","TOO_EARLY","TOO_MANY_REQUESTS","UNAUTHORIZED","UNAVAILABLE_FOR_LEGAL_REASONS","UNPROCESSABLE_ENTITY","UNSUPPORTED_MEDIA_TYPE","UPGRADE_REQUIRED","URI_TOO_LONG","USE_PROXY","VARIANT_ALSO_NEGOTIATES"]},"view":{"$ref":"#/components/schemas/View"},"viewName":{"type":"string"}}},"Province":{"title":"Province","type":"object","properties":{"countryCode":{"type":"string"},"provinceCode":{"type":"string"},"provinceName":{"type":"string"}}},"ResponseDTO«CustomerAddress»":{"title":"ResponseDTO«CustomerAddress»","type":"object","properties":{"data":{"$ref":"#/components/schemas/CustomerAddress"},"errors":{"uniqueItems":true,"type":"array","items":{"type":"string"}}}},"ResponseDTO«List«Country»»":{"title":"ResponseDTO«List«Country»»","type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/Country"}},"errors":{"uniqueItems":true,"type":"array","items":{"type":"string"}}}},"ResponseDTO«List«CustomerAddress»»":{"title":"ResponseDTO«List«CustomerAddress»»","type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/CustomerAddress"}},"errors":{"uniqueItems":true,"type":"array","items":{"type":"string"}}}},"ResponseDTO«List«Province»»":{"title":"ResponseDTO«List«Province»»","type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/Province"}},"errors":{"uniqueItems":true,"type":"array","items":{"type":"string"}}}},"ResponseDTO«int»":{"title":"ResponseDTO«int»","type":"object","properties":{"data":{"type":"integer","format":"int32"},"errors":{"uniqueItems":true,"type":"array","items":{"type":"string"}}}},"View":{"title":"View","type":"object","properties":{"contentType":{"type":"string"}}}}}}